requirements.txt
arrow==1.3.0
asgiref==3.9.1
binaryornot==0.4.4
boto3==1.40.44
botocore==1.40.44
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
Django==5.2
django-browser-reload==1.18.0
django-ckeditor-5==0.2.18
django-storages==1.14.6
django-tailwind==4.2.0
django-widget-tweaks==1.5.0
honcho==2.0.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pillow==11.3.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20250822
tzdata==2025.2
urllib3==2.5.0
